+++ src/dune_rules/dune_env.ml
@@ -15,6 +15,13 @@ module Stanza = struct
     and+ cxx = Ordered_set_lang.Unexpanded.field "cxx_flags" ?check in
     Foreign_language.Dict.make ~c ~cxx
 
+  let link_flags ~since =
+    let check =
+      Option.map since ~f:(fun since ->
+          Dune_lang.Syntax.since Stanza.syntax since)
+    in
+    Ordered_set_lang.Unexpanded.field "link_flags" ?check
+
   let menhir_flags ~since =
     let check =
       Option.map since ~f:(fun since ->
@@ -74,6 +81,7 @@ module Stanza = struct
   type config =
     { flags : Ocaml_flags.Spec.t
     ; foreign_flags : Ordered_set_lang.Unexpanded.t Foreign_language.Dict.t
+    ; link_flags : Ordered_set_lang.Unexpanded.t
     ; env_vars : Env.t
     ; binaries : File_binding.Unexpanded.t list
     ; inline_tests : Inline_tests.t option
@@ -87,6 +95,7 @@ module Stanza = struct
   let equal_config
       { flags
       ; foreign_flags
+      ; link_flags
       ; env_vars
       ; binaries
       ; inline_tests
@@ -99,6 +108,7 @@ module Stanza = struct
     Ocaml_flags.Spec.equal flags t.flags
     && Foreign_language.Dict.equal Ordered_set_lang.Unexpanded.equal
          foreign_flags t.foreign_flags
+    && Ordered_set_lang.Unexpanded.equal link_flags t.link_flags
     && Env.equal env_vars t.env_vars
     && List.equal File_binding.Unexpanded.equal binaries t.binaries
     && Option.equal Inline_tests.equal inline_tests t.inline_tests
@@ -114,6 +124,7 @@ module Stanza = struct
     { flags = Ocaml_flags.Spec.standard
     ; foreign_flags =
         Foreign_language.Dict.make_both Ordered_set_lang.Unexpanded.standard
+    ; link_flags = Ordered_set_lang.Unexpanded.standard
     ; env_vars = Env.empty
     ; binaries = []
     ; inline_tests = None
@@ -181,6 +192,7 @@ module Stanza = struct
   let config =
     let+ flags = Ocaml_flags.Spec.decode
     and+ foreign_flags = foreign_flags ~since:(Some (1, 7))
+    and+ link_flags = link_flags ~since:(Some (3, 0))
     and+ env_vars = env_vars_field
     and+ binaries =
       field ~default:[] "binaries"
@@ -194,6 +206,7 @@ module Stanza = struct
     and+ format_config = Format_config.field ~since:(2, 8) in
     { flags
     ; foreign_flags
+    ; link_flags
     ; env_vars
     ; binaries
     ; inline_tests